ir_sample/cashflow-scanner/requirements.txt
google-generativeai
pydantic
pymupdf
